requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decord==0.6.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
filelock==3.18.0
flash_attn==2.8.0.post2
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
Gpt_Client==1.28
h11==0.14.0
hf-xet==1.1.5
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.33.1
idna==3.10
imgcat==0.6.0
Jinja2==3.1.6
jiter==0.10.0
kaggle==1.7.4.5
kiwisolver==1.4.8
logzero==1.7.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.57.1
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
safetensors==0.5.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
text-unidecode==1.3
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.1
